app/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.23
alembic==1.13.1
pydantic-settings==2.2.1
Pillow==10.4.0
pytest==8.3.3
httpx==0.27.2
pytesseract==0.3.10
openai==2.7.2
pytz==2024.1
